* Replace placeholder sound code in cutscenes with "GameSound". * This removes "lsound" since it is no longer required. * Fixes a few small mistakes in the RE iMuse digital sound found by actually using it. * Changed the imMidiFlag from 0x80000000u to 0x8000000000000000u to account for 64-bit pointers. The midi mask was changed accordingly. * GameSound cleanup.

+++ TheForceEngine/TFE_Jedi/Sound/gameSound.h
@@ -34,6 +34,8 @@ enum GameSoundType
 	digitalSound
 };
 
+// Structure padding matters due to the hacky way iMuse gets at the sound data,
+// so I promoted 16-bit members to 32-bit to simplify.
 struct GameSound
 {
 	u32 resType;
@@ -41,13 +43,13 @@ struct GameSound
 
 	GameSound* next;
 	GameSoundType type;
-	s16 id;
+	s32 id;
 
-	s16 flags;
-	s16 volume;
+	s32 flags;
+	s32 volume;
 
-	s16 var1;
-	s16 var2;
+	s32 var1;
+	s32 var2;
 	u8* varptr;
 	u8* varhdl;
